+++ benches/bimodal_ke.rs
@@ -9,8 +9,8 @@ fn create_equation() -> equation::ODE {
             fetch_params!(p, ke, _v);
             dx[0] = -ke * x[0] + rateiv[0];
         },
-        |_p| lag! {},
-        |_p| fa! {},
+        |_p, _t, _cov| lag! {},
+        |_p, _t, _cov| fa! {},
         |_p, _t, _cov, _x| {},
         |x, p, _t, _cov, y| {
             fetch_params!(p, _ke, v);
@@ -22,13 +22,20 @@ fn create_equation() -> equation::ODE {
 
 fn setup_simulation() -> Result<(Settings, equation::ODE, data::Data)> {
     let params = Parameters::new()
-        .add("ke", 0.001, 3.0, false)
-        .add("v", 25.0, 250.0, false);
+        .add("ke", 0.001, 3.0)
+        .add("v", 25.0, 250.0);
+
+    let ems = ErrorModels::new()
+        .add(
+            0,
+            ErrorModel::additive(ErrorPoly::new(0.0, 0.5, 0.0, 0.0), 0.0),
+        )
+        .unwrap();
 
     let mut settings = Settings::builder()
         .set_algorithm(Algorithm::NPAG)
         .set_parameters(params)
-        .set_error_model(ErrorType::Additive, 0.0, (0.0, 0.5, 0.0, 0.0))
+        .set_error_models(ems)
         .build();
 
     settings.set_cycles(1000);