+++ examples/iov/main.rs
@@ -15,8 +15,8 @@ fn main() -> Result<()> {
             fetch_params!(p, _ke0);
             d[1] = 0.1;
         },
-        |_p| lag! {},
-        |_p| fa! {},
+        |_p, _t, _cov| lag! {},
+        |_p, _t, _cov| fa! {},
         |p, _t, _cov, x| {
             fetch_params!(p, ke0);
             x[1] = ke0;
@@ -29,16 +29,17 @@ fn main() -> Result<()> {
         10000,
     );
 
-    let params = Parameters::new().add("ke0", 0.001, 2.0, false);
+    let params = Parameters::new().add("ke0", 0.001, 2.0);
+
+    let ems = ErrorModels::new().add(
+        0,
+        ErrorModel::additive(ErrorPoly::new(0.0, 0.0, 0.0, 0.0), 0.0000757575757576),
+    )?;
 
     let mut settings = Settings::builder()
         .set_algorithm(Algorithm::NPAG)
         .set_parameters(params)
-        .set_error_model(
-            ErrorType::Additive,
-            0.0000757575757576,
-            (0.0, 0.0, 0.0, 0.0),
-        )
+        .set_error_models(ems)
         .build();
 
     settings.set_cycles(100000);