+++ examples/theophylline/main.rs
@@ -8,8 +8,8 @@ fn main() {
     //         dx[0] = -ka * x[0];
     //         dx[1] = ka * x[0] - ke * x[1];
     //     },
-    //     |_p| lag! {},
-    //     |_p| fa! {},
+    //     |_p, _t, _cov| lag! {},
+    //     |_p, _t, _cov| fa! {},
     //     |_p, _t, _cov, _x| {},
     //     |x, p, _t, _cov, y| {
     //         fetch_params!(p, _ka, _ke, v);
@@ -20,8 +20,8 @@ fn main() {
     let eq = equation::Analytical::new(
         one_compartment_with_absorption,
         |_p, _t, _cov| {},
-        |_p| lag! {},
-        |_p| fa! {},
+        |_p, _t, _cov| lag! {},
+        |_p, _t, _cov| fa! {},
         |_p, _t, _cov, _x| {},
         |x, p, _t, _cov, y| {
             fetch_params!(p, _ka, _ke, v);
@@ -31,14 +31,21 @@ fn main() {
     );
 
     let params = Parameters::new()
-        .add("ka", 0.001, 3.0, false)
-        .add("ke", 0.001, 3.0, false)
-        .add("v", 0.001, 50.0, false);
+        .add("ka", 0.001, 3.0)
+        .add("ke", 0.001, 3.0)
+        .add("v", 0.001, 50.0);
+
+    let ems = ErrorModels::new()
+        .add(
+            0,
+            ErrorModel::proportional(ErrorPoly::new(0.1, 0.1, 0.0, 0.0), 2.0),
+        )
+        .unwrap();
 
     let mut settings = Settings::builder()
         .set_algorithm(Algorithm::NPAG)
         .set_parameters(params)
-        .set_error_model(ErrorType::Proportional, 2.0, (0.1, 0.1, 0.0, 0.0))
+        .set_error_models(ems)
         .build();
 
     settings.initialize_logs().unwrap();