+++ src/algorithms/npod.rs
@@ -1,4 +1,5 @@
 use crate::{
+    algorithms::Status,
     prelude::{
         algorithms::Algorithms,
         routines::{
@@ -14,16 +15,13 @@ use crate::{
 };
 use anyhow::bail;
 use anyhow::Result;
+use faer::Col;
 use faer_ext::IntoNdarray;
+use pharmsol::{prelude::ErrorModel, ErrorModels};
 use pharmsol::{
-    prelude::{
-        data::{Data, ErrorModel},
-        simulator::Equation,
-    },
+    prelude::{data::Data, simulator::Equation},
     Subject,
 };
-
-use faer::Col;
 
 use ndarray::{
     parallel::prelude::{IntoParallelRefMutIterator, ParallelIterator},
@@ -37,7 +35,6 @@ const THETA_D: f64 = 1e-4;
 
 pub struct NPOD<E: Equation> {
     equation: E,
-    ranges: Vec<(f64, f64)>,
     psi: Psi,
     theta: Theta,
     lambda: Col<f64>,
@@ -45,9 +42,10 @@ pub struct NPOD<E: Equation> {
     last_objf: f64,
     objf: f64,
     cycle: usize,
-    gamma_delta: f64,
-    error_model: ErrorModel,
+    gamma_delta: Vec<f64>,
+    error_models: ErrorModels,
     converged: bool,
+    status: Status,
     cycle_log: CycleLog,
     data: Data,
     settings: Settings,
@@ -57,7 +55,6 @@ impl<E: Equation> Algorithms<E> for NPOD<E> {
     fn new(settings: Settings, equation: E, data: Data) -> Result<Box<Self>, anyhow::Error> {
         Ok(Box::new(Self {
             equation,
-            ranges: settings.parameters().ranges(),
             psi: Psi::new(),
             theta: Theta::new(),
             lambda: Col::zeros(0),
@@ -65,9 +62,10 @@ impl<E: Equation> Algorithms<E> for NPOD<E> {
             last_objf: -1e30,
             objf: f64::NEG_INFINITY,
             cycle: 0,
-            gamma_delta: 0.1,
-            error_model: settings.error().clone().into(),
+            gamma_delta: vec![0.1; settings.errormodels().len()],
+            error_models: settings.errormodels().clone().into(),
             converged: false,
+            status: Status::Starting,
             cycle_log: CycleLog::new(),
             settings,
             data,
@@ -82,7 +80,7 @@ impl<E: Equation> Algorithms<E> for NPOD<E> {
             self.w.clone(),
             -2. * self.objf,
             self.cycle,
-            self.converged,
+            self.status.clone(),
             self.settings.clone(),
             self.cycle_log.clone(),
         )
@@ -117,7 +115,7 @@ impl<E: Equation> Algorithms<E> for NPOD<E> {
         self.theta = theta;
     }
 
-    fn get_theta(&self) -> &Theta {
+    fn theta(&self) -> &Theta {
         &self.theta
     }
 
@@ -129,22 +127,33 @@ impl<E: Equation> Algorithms<E> for NPOD<E> {
         self.objf
     }
 
+    fn set_status(&mut self, status: Status) {
+        self.status = status;
+    }
+
+    fn status(&self) -> &Status {
+        &self.status
+    }
+
     fn convergence_evaluation(&mut self) {
         if (self.last_objf - self.objf).abs() <= THETA_F {
             tracing::info!("Objective function convergence reached");
             self.converged = true;
+            self.status = Status::Converged;
         }
 
         // Stop if we have reached maximum number of cycles
         if self.cycle >= self.settings.config().cycles {
             tracing::warn!("Maximum number of cycles reached");
             self.converged = true;
+            self.status = Status::MaxCycles;
         }
 
         // Stop if stopfile exists
         if std::path::Path::new("stop").exists() {
             tracing::warn!("Stopfile detected - breaking");
             self.converged = true;
+            self.status = Status::ManualStop;
         }
 
         // Create state object
@@ -154,8 +163,8 @@ impl<E: Equation> Algorithms<E> for NPOD<E> {
             delta_objf: (self.last_objf - self.objf).abs(),
             nspp: self.theta.nspp(),
             theta: self.theta.clone(),
-            gamlam: self.error_model.scalar(),
-            converged: self.converged,
+            error_models: self.error_models.clone(),
+            status: self.status.clone(),
         };
 
         // Write cycle log
@@ -168,7 +177,7 @@ impl<E: Equation> Algorithms<E> for NPOD<E> {
     }
 
     fn evaluation(&mut self) -> Result<()> {
-        let error_model: ErrorModel = self.error_model.clone();
+        let error_model: ErrorModels = self.error_models.clone();
 
         self.psi = calculate_psi(
             &self.equation,
@@ -252,72 +261,95 @@ impl<E: Equation> Algorithms<E> for NPOD<E> {
     }
 
     fn optimizations(&mut self) -> Result<()> {
-        // Gam/Lam optimization
-        // TODO: Move this to e.g. /evaluation/error.rs
-        let gamma_up = self.error_model.scalar() * (1.0 + self.gamma_delta);
-        let gamma_down = self.error_model.scalar() / (1.0 + self.gamma_delta);
+        self.error_models
+            .clone()
+            .iter_mut()
+            .filter_map(|(outeq, em)| match em {
+                ErrorModel::None => None,
+                _ => Some((outeq, em)),
+            })
+            .try_for_each(|(outeq, em)| -> Result<()> {
+                // OPTIMIZATION
 
-        let mut error_model_up: ErrorModel = self.error_model.clone();
-        error_model_up.set_scalar(gamma_up);
+                let gamma_up = em.scalar()? * (1.0 + self.gamma_delta[outeq]);
+                let gamma_down = em.scalar()? / (1.0 + self.gamma_delta[outeq]);
 
-        let mut error_model_down: ErrorModel = self.error_model.clone();
-        error_model_down.set_scalar(gamma_down);
+                let mut error_model_up = self.error_models.clone();
+                error_model_up.set_scalar(outeq, gamma_up)?;
 
-        let psi_up = calculate_psi(
-            &self.equation,
-            &self.data,
-            &self.theta,
-            &error_model_up,
-            false,
-            true,
-        )?;
-        let psi_down = calculate_psi(
-            &self.equation,
-            &self.data,
-            &self.theta,
-            &error_model_down,
-            false,
-            true,
-        )?;
+                let mut error_model_down = self.error_models.clone();
+                error_model_down.set_scalar(outeq, gamma_down)?;
 
-        let (lambda_up, objf_up) = match burke(&psi_up) {
-            Ok((lambda, objf)) => (lambda, objf),
-            Err(err) => {
-                return Err(anyhow::anyhow!("Error in IPM: {:?}", err));
-            }
-        };
-        let (lambda_down, objf_down) = match burke(&psi_down) {
-            Ok((lambda, objf)) => (lambda, objf),
-            Err(err) => {
-                return Err(anyhow::anyhow!("Error in IPM: {:?}", err));
-            }
-        };
+                let psi_up = calculate_psi(
+                    &self.equation,
+                    &self.data,
+                    &self.theta,
+                    &error_model_up,
+                    false,
+                    true,
+                )?;
+                let psi_down = calculate_psi(
+                    &self.equation,
+                    &self.data,
+                    &self.theta,
+                    &error_model_down,
+                    false,
+                    true,
+                )?;
 
-        if objf_up > self.objf {
-            self.error_model.set_scalar(gamma_up);
-            self.objf = objf_up;
-            self.gamma_delta *= 4.;
-            self.lambda = lambda_up;
-            self.psi = psi_up;
-        }
-        if objf_down > self.objf {
-            self.error_model.set_scalar(gamma_down);
-            self.objf = objf_down;
-            self.gamma_delta *= 4.;
-            self.lambda = lambda_down;
-            self.psi = psi_down;
-        }
-        self.gamma_delta *= 0.5;
-        if self.gamma_delta <= 0.01 {
-            self.gamma_delta = 0.1;
-        }
+                let (lambda_up, objf_up) = match burke(&psi_up) {
+                    Ok((lambda, objf)) => (lambda, objf),
+                    Err(err) => {
+                        //todo: write out report
+                        return Err(anyhow::anyhow!("Error in IPM during optim: {:?}", err));
+                    }
+                };
+                let (lambda_down, objf_down) = match burke(&psi_down) {
+                    Ok((lambda, objf)) => (lambda, objf),
+                    Err(err) => {
+                        //todo: write out report
+                        //panic!("Error in IPM: {:?}", err);
+                        return Err(anyhow::anyhow!("Error in IPM during optim: {:?}", err));
+                        //(Array1::zeros(1), f64::NEG_INFINITY)
+                    }
+                };
+                if objf_up > self.objf {
+                    self.error_models.set_scalar(outeq, gamma_up)?;
+                    self.objf = objf_up;
+                    self.gamma_delta[outeq] *= 4.;
+                    self.lambda = lambda_up;
+                    self.psi = psi_up;
+                }
+                if objf_down > self.objf {
+                    self.error_models.set_scalar(outeq, gamma_down)?;
+                    self.objf = objf_down;
+                    self.gamma_delta[outeq] *= 4.;
+                    self.lambda = lambda_down;
+                    self.psi = psi_down;
+                }
+                self.gamma_delta[outeq] *= 0.5;
+                if self.gamma_delta[outeq] <= 0.01 {
+                    self.gamma_delta[outeq] = 0.1;
+                }
+                Ok(())
+            })?;
+
         Ok(())
     }
 
     fn logs(&self) {
         tracing::info!("Objective function = {:.4}", -2.0 * self.objf);
         tracing::debug!("Support points: {}", self.theta.nspp());
-        tracing::debug!("Gamma = {:.16}", self.error_model.scalar());
+        self.error_models.iter().for_each(|(outeq, em)| {
+            if ErrorModel::None == *em {
+                return;
+            }
+            tracing::debug!(
+                "Error model for outeq {}: {:.16}",
+                outeq,
+                em.scalar().unwrap_or_default()
+            );
+        });
         // Increasing objf signals instability or model misspecification.
         if self.last_objf > self.objf + 1e-4 {
             tracing::warn!(
@@ -336,7 +368,7 @@ impl<E: Equation> Algorithms<E> for NPOD<E> {
         let pyl = psi.dot(&w);
 
         // Add new point to theta based on the optimization of the D function
-        let error_model: ErrorModel = self.error_model.clone();
+        let error_model: ErrorModels = self.error_models.clone();
 
         let mut candididate_points: Vec<Array1<f64>> = Vec::default();
         for spp in self.theta.matrix().row_iter() {
@@ -355,8 +387,7 @@ impl<E: Equation> Algorithms<E> for NPOD<E> {
             // re-define a new optimization
         });
         for cp in candididate_points {
-            self.theta
-                .suggest_point(cp.to_vec().as_slice(), THETA_D, &self.ranges);
+            self.theta.suggest_point(cp.to_vec().as_slice(), THETA_D);
         }
         Ok(())
     }