+++ CHANGELOG.md
@@ -7,6 +7,44 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.1](https://github.com/LAPKB/PMcore/compare/v0.13.0...v0.13.1) - 2025-07-09
+
+### Added
+
+- support for pharmsol 0.13.1 ([#153](https://github.com/LAPKB/PMcore/pull/153))
+
+## [0.13.0](https://github.com/LAPKB/PMcore/compare/v0.12.1...v0.13.0) - 2025-06-25
+
+### Added
+
+- More informative status for algorithm and stop reason ([#138](https://github.com/LAPKB/PMcore/pull/138))
+
+### Fixed
+
+- Add gamma/lambda for each output equation to cycle.csv ([#147](https://github.com/LAPKB/PMcore/pull/147))
+
+### Other
+
+- Update Cargo.toml ([#150](https://github.com/LAPKB/PMcore/pull/150))
+- Add validation of parameters ([#136](https://github.com/LAPKB/PMcore/pull/136))
+
+## [0.12.1](https://github.com/LAPKB/PMcore/compare/v0.12.0...v0.12.1) - 2025-06-19
+
+### Other
+
+- Update Cargo.toml ([#145](https://github.com/LAPKB/PMcore/pull/145))
+- Use pharmsol 0.11.0
+
+## [0.12.0](https://github.com/LAPKB/PMcore/compare/v0.11.0...v0.12.0) - 2025-06-11
+
+### Added
+
+- Deprecate fixed but unknown ([#118](https://github.com/LAPKB/PMcore/pull/118))
+
+### Other
+
+- Error model ([#139](https://github.com/LAPKB/PMcore/pull/139))
+
 ## [0.10.0](https://github.com/LAPKB/PMcore/compare/v0.9.0...v0.10.0) - 2025-03-28
 
 ### Added